move crawler to script
